services/finance_worker/pyproject.toml
[project]
name = "finance-worker"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "alpaca-py>=0.42.2",
    "pika>=1.3.2",
]

[dependency-groups]
dev = []
